requirements.txt
flask
opencv-python-headless
numpy
rknn-toolkit-lite2
pyserial
